mv normalize-path to new file && remove extra ternary
